Strange wording in [temp.constr.constr]/2 Note.
"The satisfaction of constraints on class templates, alias templates,
and variable templates is required when referring a template specialization
(\ref{temp.names})." is worded oddly. One does not "refer" a template specialization. I believe it should be "when referencing a template specialization," or perhaps "when a template specialization is referenced."